% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrpthdrs-package.R
\docType{package}
\name{scrpthdrs-package}
\alias{scrpthdrs}
\alias{scrpthdrs-package}
\title{Script Headers package}
\description{
Insert comment headers in your R script files opened in RStudio to include basic meta data.
Learn more about this package in the vignette "insert_script_headers":
\code{vignette(insert_script_headers, package = "scrpthdrs")}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://philippbaessler.github.io/scrpthdrs/}
}

}
\author{
\strong{Maintainer}: Philipp Baessler \email{dev@philipb.com}

}
\keyword{internal}
