% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrpthdrs-package.R
\docType{package}
\name{scrpthdrs-package}
\alias{scrpthdrs}
\alias{scrpthdrs-package}
\title{scrpthdrs: Script Headers}
\description{
Provide functionality to easily insert a comment section at the top of a R script file with information provided by the user and basic meta data. This package comprises an RStudio addin with a basic GUI.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://philippbaessler.github.io/scrpthdrs/}
}

}
\author{
\strong{Maintainer}: Philipp Baessler \email{dev@philipb.com}

}
\keyword{internal}
